# openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Local Access API
  description: 'API that provides read access to the users OS X local file system'
  contact:
    email: ordog.ivett@gmail.com
servers:
  - url: http://localhost:5004
paths:
  /listDir:
    get:
      operationId: listDirectory
      summary: List directory contents and file information
      parameters:
        - name: path
          in: query
          description: The directory path
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: The page number
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: The number of results per page
          required: false
          schema:
            type: integer
            default: 100
            maximum: 100
        - name: order
          in: query
          description: The order of the files (by name or size)
          required: false
          schema:
            type: string
            enum: [name, size]
            default: name
        - name: direction
          in: query
          description: The direction of the order (ascending or descending)
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: A list of files and their information
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      description: The file name
                    size:
                      type: integer
                      format: int64
                      description: The file size
                    modified:
                      type: string
                      format: date-time
                      description: The last modification time
        '400':
          description: Error message
  /getFileInfo:
    get:
      operationId: getFileInfo
      summary: Get file information
      parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
          description: The file path
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  size:
                    type: string
                  modified:
                    type: string
  /getFileContent:
    get:
      operationId: getFileContent
      summary: Get file content
      parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
          description: The file path
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          required: false
          description: The offset to start reading from, in characters
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string